+++ src/js/inject/LaunchFilter.js
@@ -1,12 +1,9 @@
-if (document.readyState == "interactive") {
-    console.log('DOM interactive');
-    TEMP = new FILTER();
-} else if (document.readyState == "complete") {
-    console.log('DOM complete');
+var TEMP; // Actually not finished
+
+if (document.readyState == "interactive" || document.readyState == "complete") {
     TEMP = new FILTER();
 } else {
     document.addEventListener('DOMContentLoaded', function() {
-        console.log('DOM loaded');
         TEMP = new FILTER();
     });
 }